requirements.txt
jishaku
aiosqlite
git+https://github.com/ftpskid/button_paginator.git
requests
wavelink
openai
psutil
pytz
orjson
forex-python
aiohttp
aiosignal
astunparse
async-timeout
asyncpg
attrs
backoff
braceexpand
certifi
charset-normalizer
click
colorama
contourpy
cycler
emoji
fonttools
frozenlist
grapheme
idna
import-expression
kiwisolver
line-profiler
matplotlib
multidict
numpy
serpapi
packaging
Pillow
psutil
pyparsing
python-dateutil
pytz
redis
requests
six
spotify
spotipy
tabulate
topggpy
typing-extensions
urllib3
yarl
